requirements.txt
torch
torchvision
s3fs
mlflow
pillow